assistants-ai-workers/qa_robot/requirements.txt
google-generativeai
streamlit
pillow
pdf2image
pandas
openpyxl
pytesseract 
PyMuPDF
python-docx
langdetect
langcodes
language_data
thefuzz
openpyxl
tqdm
openai
unidecode
underthesea
